Fix clang warnings.
* src/ftstring.c (Process_Event), src/ftview.c (Process_Event): Add
cast.

+++ src/ftstring.c
@@ -445,7 +445,7 @@
 
     if ( event->key >= '1' && event->key < '1' + N_RENDER_MODES )
     {
-      status.render_mode = event->key - '1';
+      status.render_mode = (int)( event->key - '1' );
       event_render_mode_change( 0 );
 
       return ret;
